+++ examples/tailwindcss/components/features.tsx
@@ -1,79 +1,27 @@
-import { LinkIcon } from '@chakra-ui/icons';
-import {
-  Box,
-  Heading,
-  Icon,
-  Link,
-  Stack,
-  Text,
-  useColorModeValue
-} from '@chakra-ui/react';
-import { FeatureProps } from './types';
+import { LinkIcon } from '@heroicons/react/24/outline'
+import { FeatureProps } from './types'
 
 export const Product = ({ title, text, href }: FeatureProps) => {
   return (
-    <Link href={href} target="_blank" _hover={{ textDecoration: 'none' }}>
-      <Stack
-        h="full"
-        minH={36}
-        p={5}
-        spacing={2.5}
-        justifyContent="center"
-        borderRadius={5}
-        boxShadow={useColorModeValue(
-          '0 2px 5px #ccc',
-          '0 1px 3px #727272, 0 2px 12px -2px #2f2f2f'
-        )}
-        _hover={{
-          color: useColorModeValue('purple.600', 'purple.300'),
-          boxShadow: useColorModeValue(
-            '0 2px 5px #bca5e9',
-            '0 0 3px rgba(150, 75, 213, 0.8), 0 3px 8px -2px rgba(175, 89, 246, 0.9)'
-          )
-        }}
-      >
-        <Heading fontSize="xl">{title}&ensp;&rarr;</Heading>
-        <Text>{text}</Text>
-      </Stack>
-    </Link>
-  );
-};
+    <a href={href} target="_blank" rel="noopener noreferrer">
+      <div className="h-full [min-height:9rem] p-5 justify-center border rounded-lg border-black/10 transition duration-75 hover:shadow-xl dark:border-white/10 hover:bg-black/10 dark:hover:bg-white/10">
+        <h3 className="text-lg font-medium">{title}&ensp;&rarr;</h3>
+        <p className="mt-1 text-black/75 dark:text-white/75">{text}</p>
+      </div>
+    </a>
+  )
+}
 
 export const Dependency = ({ title, text, href }: FeatureProps) => {
   return (
-    <Link href={href} target="_blank" _hover={{ textDecoration: 'none' }}>
-      <Stack
-        isInline={true}
-        key={title}
-        spacing={3}
-        h="full"
-        p={4}
-        justifyContent="center"
-        borderRadius="md"
-        border="1px solid"
-        borderColor={useColorModeValue('blackAlpha.200', 'whiteAlpha.100')}
-        _hover={{
-          boxShadow: useColorModeValue(
-            '0 2px 5px #ccc',
-            '0 1px 3px #727272, 0 2px 12px -2px #2f2f2f'
-          )
-        }}
-      >
-        <Box color={useColorModeValue('primary.500', 'primary.200')}>
-          <Icon as={LinkIcon} />
-        </Box>
-        <Stack spacing={1}>
-          <Text fontSize="lg" fontWeight="semibold">
-            {title}
-          </Text>
-          <Text
-            lineHeight="short"
-            color={useColorModeValue('blackAlpha.700', 'whiteAlpha.700')}
-          >
-            {text}
-          </Text>
-        </Stack>
-      </Stack>
-    </Link>
-  );
-};
+    <a href={href} target="_blank" rel="noopener noreferrer">
+      <div className="flex flex-row items-start justify-center h-full p-5 space-x-2 transition duration-75 border rounded-lg border-black/10 hover:shadow-xl dark:border-white/10 hover:bg-black/10 dark:hover:bg-white/10">
+        <LinkIcon className="flex-shrink-0 w-5 h-5 mt-1 text-black/75 dark:text-white/75" />
+        <div>
+          <h3 className="text-lg font-medium">{title}&ensp;&rarr;</h3>
+          <p className="mt-1 text-black/75 dark:text-white/75">{text}</p>
+        </div>
+      </div>
+    </a>
+  )
+}