+++ examples/tailwindcss/components/react/views/index.ts
@@ -1,0 +1,6 @@
+export * from './Connected';
+export * from './Connecting';
+export * from './Error';
+export * from './NotExist';
+export * from './QRCode';
+export * from './WalletList';