+++ templates/connect-multi-chain/pages/_app.tsx
@@ -5,7 +5,7 @@ import { ChakraProvider } from '@chakra-ui/react';
 import { defaultTheme } from '../config';
 import { wallets } from '@cosmos-kit/keplr';
 import { SignerOptions } from '@cosmos-kit/core';
-import { chains } from 'chain-registry';
+import { chains, assets } from 'chain-registry';
 
 function CreateCosmosApp({ Component, pageProps }: AppProps) {
   const signerOptions: SignerOptions = {
@@ -18,6 +18,7 @@ function CreateCosmosApp({ Component, pageProps }: AppProps) {
     <ChakraProvider theme={defaultTheme}>
       <WalletProvider
         chains={chains}
+        assetLists={assets}
         wallets={wallets}
         signerOptions={signerOptions}
       >