+++ packages/create-cosmwasm-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmwasm-app@1.0.0...create-cosmwasm-app@1.0.1) (2022-10-01)
+
+**Note:** Version bump only for package create-cosmwasm-app
+
+
+
+
+
+# [1.0.0](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmwasm-app@0.5.1...create-cosmwasm-app@1.0.0) (2022-10-01)
+
+**Note:** Version bump only for package create-cosmwasm-app
+
+
+
+
+
 ## [0.5.1](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmwasm-app@0.5.0...create-cosmwasm-app@0.5.1) (2022-09-30)
 
 **Note:** Version bump only for package create-cosmwasm-app